refactor: extract IndexServer into CodeSearchService
